backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0

